requirements.txt
django==2.0
Pillow
django-smart-selects